+++ CHANGELOG.md
@@ -8,10 +8,13 @@
 
 ## master branch (latest changes not released yet)
 
+- improved notebooks
+- added C and L components from phidl
+- print(component) returns more info (similar to phidl)
+
 ## 1.4.3 2020-10-07
 
 - clean metadata dict recursively
-
 
 ## 1.4.2 2020-10-07